*** Important changes in this version ***
This version adds the option of upsampling along the frequency dimension
and changes some of the default settings. 

The following new properties of the dbt object control upsampling:

1. upsampleFx increases frequency sampling by a factor of (value + 1). In
   otherwords, upsampleFx = 0 implies no upsampling while upsampleFx = 1
   implies 2x upsampling. 

   upsampleFx can take a non-integer value, however signal reconstruction
   through the inverse transform is not implemented for this case, and an error will
   be thrown if the signal method is called. 

2. Upsampling in time is still controlled through the fftpad property. In addition,
   the keyword 'upsampleTx' is accepted by the constructor, which is synonymous 
   with the 'fftpad' keyword.  

3. The dbt constructor can take the keyword "upsample", followed by a value,
   in which case both time and frequency will be upsampled by a factor of value.
   In other words, 

            obj = dbt(...,'upsample',3)

   returns a dbt object in which both time and frequency are upsampled by a 
   a factor of 3. Note this differs from upsampleFx and fftpad which upsample
   by value + 1. 


The following default settings are new or changed:

1. UpsampleFx is 0 by default

2. centerDC is now true by default. This means there is no default signal
   upsampling as before. 

3. remodphase is false by default.



Miscellaneous changes: 

Scaling of the DC band has been changed. Previously, the DC component of the
fft of the entire signal was scaled by 1/sqrt(2) going into the DBT to compensate
for the sqrt(2) scaling applied to all bands. This didn't make a lot of sense. 
A more sensible option would be to leave the DC band unscaled while scaling 
other bands by sqrt(2) to preserve total signal energy. That option is not 
what has been done in this revision. In the current version the DC band 
is instead scaled like the other bands so that power spectral density remains 
comparable across bands. 


 
   

git-svn-id: https://saccade.neurosurgery.uiowa.edu/svn/KovachToolbox@621 4e33aaf1-3bdf-43e8-baf7-e7e736085cc8